Refactor: Rename `type.ts` to `types.ts` and update imports
This commit renames the file `type.ts` to `types.ts` to better reflect its contents, which include multiple type definitions. All corresponding imports in the project have been updated to reflect this change, ensuring consistency and clarity in type usage across the codebase.
